reorganize for split cache and topic configs
