+++ src/cache/config/mod.rs
@@ -1,0 +1,4 @@
+/// Configuration for the Momento cache client.
+pub mod configuration;
+/// Pre-built configurations for the Momento cache client.
+pub mod configurations;
