+++ src/config/mod.rs
@@ -1,7 +1,3 @@
-/// Configuration for the Momento cache client.
-pub mod configuration;
-/// Pre-built configurations for the Momento cache client.
-pub mod configurations;
 /// Low-level gRPC settings for communicating with Momento.
 pub mod grpc_configuration;
 /// Low-level settings for communicating with Momento.
