+++ src/topics/mod.rs
@@ -1,0 +1,3 @@
+mod topic_client;
+
+pub use topic_client::TopicClient;
